requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cfgv==3.4.0
click==8.3.0
distlib==0.4.0
fastapi==0.116.2
filelock==3.19.1
flake8==7.3.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.14
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-env==1.1.5
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
virtualenv==20.34.0
